chore(audit): drive lint, gosec and documentation to zero defects
Five rounds of a builder/critic loop over the whole Go source, the 30
markdown documents and the 11 Mermaid diagrams.

Gate, measured rather than asserted:

  golangci-lint run ./...                     152 -> 0
  gosec ./...                                  19 -> 0
  gofmt -l .                                    1 -> 0
  go build / go vet / go test -count=1 ./...   green throughout

Reached without loosening the tools. .golangci.yml changed in comments
only across all six passes: no exclusion widened, no threshold moved, no
rule added or removed. //nolint stays at 4, in the same four files.
#nosec falls 22 -> 13 because suppression sites were removed rather than
annotated: conversions deleted where Go already accepts a signed shift
count, and a real bounds guard added inside bigfft.valueSize, after which
gosec accepts the cast with no annotation at all.

Tests gained rather than lost: 877 -> 879 functions, 408 -> 410 subtests,
2291 -> 2305 assertions, zero t.Skip added. Three tests were added and
each was proven non-vacuous by mutating the code it covers:

  - bigfft.valueSize bounds, one subtest per side of the limit
  - poolAllocator.allocFermatSlice allocation steady state; the package
    had no allocation measurement on its production allocator at all
  - the cached calibration profile overriding explicit CLI flags

Documentation: false claims corrected against the source, one command per
claim. Symbol names and export status, signatures, counts, thresholds,
flags, environment variables, package edges, call orderings, branch
conditions, and cited line numbers. Every file:line reference into Go
source became a symbol anchor (121 -> 3; the three remaining are quoted
compiler output), so a refactor can no longer rot them silently.

Two behaviours are now documented and deliberately unchanged:

  - a valid cached calibration profile overrides explicit --threshold,
    --fft-threshold and --strassen-threshold, because
    calibration.LoadCachedCalibration overwrites them without consulting
    the flag set. Every document claimed the opposite. Now pinned by
    TestNewCachedProfileOverridesExplicitFlags.
  - GetDefaultProfilePath falls back to a bare relative name when
    os.UserHomeDir fails, so reads and writes land in the process CWD.

Every performance figure with no artifact behind it was removed rather
than restated. docs/audits/bench-baseline.txt is the only measurement the
tree carries; the reference table is now its medians.

Known and left in place, each documented where it belongs:
internal/fibonacci TestStateBump_PinnedAcrossCachedCalls is flaky at the
same rate on the parent commit (a prior test leaves an oversized arena in
the global state pool), and the tree does not build for 32-bit targets
(maxReasonableWords overflows a 32-bit int). -race could not run on this
host, so the concurrency conclusions here are static, not measured.

+++ internal/metrics/indicators.go
@@ -16,7 +16,7 @@ type Indicators struct {
 	// Performance
 	BitsPerSecond   float64 // bits of result produced per second
 	DigitsPerSecond float64 // estimated decimal digits produced per second
-	DoublingSteps   uint64  // number of doubling iterations ≈ log₂(n)
+	DoublingSteps   int     // number of doubling iterations ≈ log₂(n), i.e. bits.Len64(n) ∈ [0, 64]
 	StepsPerSecond  float64 // doubling steps executed per second
 
 	// Mathematical (only available after calculation completes)
@@ -40,8 +40,7 @@ var lastDigitsMod = new(big.Int).Exp(big.NewInt(10), big.NewInt(20), nil)
 func ComputeLive(n uint64, progress float64, elapsed time.Duration) *Indicators {
 	if elapsed <= 0 || progress <= 0 || n <= 1 {
 		return &Indicators{
-			// bits.Len64 returns int in [0, 64], safe to cast to uint64. #nosec G115
-			DoublingSteps: uint64(bits.Len64(n)),
+			DoublingSteps: bits.Len64(n),
 			IsEven:        n%3 == 0,
 		}
 	}
@@ -50,8 +49,7 @@ func ComputeLive(n uint64, progress float64, elapsed time.Duration) *Indicators 
 	theoreticalBits := float64(n) * log2Phi
 	estimatedBitsProduced := progress * theoreticalBits
 	estimatedDigitsProduced := estimatedBitsProduced * math.Log10(2)
-	// bits.Len64 returns int in [0, 64], safe to cast to uint64. #nosec G115
-	doublingSteps := uint64(bits.Len64(n))
+	doublingSteps := bits.Len64(n)
 	completedSteps := progress * float64(doublingSteps)
 
 	return &Indicators{
@@ -74,8 +72,7 @@ func Compute(result *big.Int, n uint64, duration time.Duration) *Indicators {
 	bitLen := result.BitLen()
 	seconds := duration.Seconds()
 	estimatedDigits := float64(bitLen) * math.Log10(2)
-	// bits.Len64 returns int in [0, 64], safe to cast to uint64. #nosec G115
-	doublingSteps := uint64(bits.Len64(n))
+	doublingSteps := bits.Len64(n)
 
 	ind := &Indicators{
 		BitsPerSecond:   float64(bitLen) / seconds,
